Two-player initial online version
In this initial version I'm only considering players the ones with ids 1
and 2. Every new connection to the server increases the id in one unit.
This means that only the first two sessions can play and if one of them
disconnects (e.g. reload the page), they won't be able to play again,
only if the server is restarted.